+++ src/main/java/salute/oneshot/domain/auth/controller/AuthViewController.java
@@ -2,10 +2,8 @@ package salute.oneshot.domain.auth.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import salute.oneshot.global.util.NonceGenerator;
 
 @Controller
 @RequestMapping("/auth")
